+++ app/static/style.css
@@ -1,0 +1,29 @@
+
+
+* {
+	margin: 0;
+	font-size: 15px;
+	line-height: 1em;
+}
+
+a {
+	text-decoration: none;
+}
+
+.main-container {
+	display: flex;
+	flex-direction: column;
+	gap: 10px;
+}
+
+.links-container {
+	display: flex;
+/*	flex-direction: column;*/
+	gap: 10px;
+}
+
+.links-container a {
+	border: 1px solid;
+	padding: 5px 10px;
+	border-radius: 5px;
+}
