+++ requirements.txt
@@ -7,3 +7,6 @@ jinja2==3.1.4
 python-multipart==0.0.9
 pandas
 websocket-client
+itsdangerous==2.2.0
+bcrypt==4.2.0
+starlette-authlib==0.3.2
